shell script remove cd, change to bash

+++ sbin/docker_remove_service.sh
@@ -1,4 +1,6 @@
-#! /bin/sh
+#! /bin/bash
 
-rm -rf $SERVICE_ROOT/$1
-rm -rf /etc/service/$1
+if [ -z $1 ] then
+  rm -rf $SERVICE_ROOT/$1
+  rm -rf /etc/service/$1
+fi
